pyproject.toml
[tool.poetry]
name = "chatyt"
version = "0.1.0"
description = "Using LLMs for question answering with youtube video transcripts as context."
authors = ["suresha97 <suresha@tcd.ie>"]

[tool.poetry.dependencies]
python = "^3.9"
youtube-data-api = "^0.0.21"
python-dotenv = "^1.0.0"
pydantic = "^2.4.2"
youtube-transcript-api = "^0.6.1"
boto3 = "^1.28.73"
langchain = "0.0.334"
tiktoken = "^0.5.1"
cohere = "^4.34"
openai = "^1.2.3"
pinecone-client = "^2.2.4"
unstructured = "^0.10.30"
pyyaml = "^6.0.1"
flask-cors = "^4.0.0"
# workaround for: https://github.com/python-poetry/poetry-plugin-export/issues/183
urllib3 = ">=1.25.4,<1.27"
werkzeug = "2.0.3"
aws-wsgi = "^0.2.7"

[tool.poetry.dev-dependencies]
pre-commit = "^3.5.0"
ruff = "^0.1.3"
mypy = "^1.6.1"
commitizen = "^3.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
